better map rendering system
in the previous commits, i was using a string of chars to store the map coordinates but after searching online i came along tiled. There was one problem though, it saves map as .tmx extension so i tried implementation it but after spending a whole day and a half i just couldn't so i switched to json parsering and it worked.